update mk commands

+++ mk
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-(( IN_NIX_DEVELOP )) || IN_NIX_DEVELOP=1 exec nix develop --command "$0" "$@"
-
 Prog=$(basename "$0")   # match what the user called
 Version=0.1
 
@@ -45,6 +43,7 @@ cmd.badges() {
 
 # cmd.code runs the current IDE
 cmd.code() {
+  (( IN_NIX_DEVELOP )) || runFlake "$0" code
   command -v cursor &>/dev/null && exec cursor .
   code .
 }
@@ -53,6 +52,8 @@ cmd.code() {
 # It parses the result from kcov's output directory.
 # The badges appear in README.md.
 cmd.cover() {
+  (( IN_NIX_DEVELOP )) || runFlake "$0" cover
+  command -v kcov &>/dev/null || { echo "kcov not found"; exit 1; }   # tool not supported on mac
   kcov --include-path tesht kcov tesht &>/dev/null
   local filenames=( $(mk.Glob kcov/tesht.*/coverage.json) )
   (( ${#filenames[*]} == 1 )) || mk.Fatal 'could not identify report file' 1
@@ -64,6 +65,8 @@ cmd.cover() {
 
 # cmd.gif creates a gif showing a sample run for README.md.
 cmd.gif() {
+  (( IN_NIX_DEVELOP )) || runFlake "$0" gif
+  command -v asciinema &>/dev/null || { echo "asciinema not found"; exit 1; }   # tool not supported on mac
   asciinema rec -c '/usr/bin/bash -c tesht' tesht.cast
   agg --speed 0.1 tesht.cast assets/tesht.gif
   rm tesht.cast
@@ -72,6 +75,7 @@ cmd.gif() {
 
 # cmd.lines determines the number of lines of source and makes a badge.
 cmd.lines() {
+  (( IN_NIX_DEVELOP )) || runFlake "$0" lines
   local lines=$(scc -f csv tesht | tail -n 1 | { IFS=, read -r language rawLines lines rest; echo $lines; })
   makeBadge "source lines" $(addCommas $lines) "#007ec6" assets/lines.svg
   echo "made source lines badge"
@@ -139,6 +143,10 @@ makeBadge() {
   <text x="150" y="14" fill="#fff" font-family="Verdana" font-size="11" text-anchor="middle">$value</text>
 </svg>
 END
+}
+
+runFlake() {
+  IN_NIX_DEVELOP=1 exec nix develop --command "$@"
 }
 
 ## globals
